Split [Conda] tests
